+++ crates/rspack/src/builder/mod.rs
@@ -1,12 +1,12 @@
+//! The Rspack compiler builder.
+
+mod builder_context;
 mod devtool;
 mod externals;
 mod target;
 
-use std::borrow::Cow;
-use std::sync::Arc;
-
+pub use builder_context::BuilderContext;
 pub use devtool::Devtool;
-use serde_json::json;
 pub use target::Targets;
 
 macro_rules! d {
@@ -33,8 +33,11 @@ macro_rules! expect {
   };
 }
 
+use std::borrow::Cow;
+use std::sync::Arc;
+
+use builder_context::BuiltinPluginOptions;
 use devtool::DevtoolFlags;
-use enum_tag::EnumTag;
 use externals::ExternalsPresets;
 use indexmap::IndexMap;
 use rspack_core::{incremental::IncrementalPasses, ModuleType};
@@ -49,8 +52,8 @@ use rspack_core::{
   JavascriptParserUrl, JsonParserOptions, LibraryName, LibraryNonUmdObject, LibraryOptions,
   LibraryType, MangleExportsOption, Mode, ModuleNoParseRules, ModuleOptions, ModuleRule,
   ModuleRuleEffect, NodeDirnameOption, NodeFilenameOption, NodeGlobalOption, NodeOption,
-  Optimization, OutputOptions, ParseOption, ParserOptions, ParserOptionsMap, PathInfo, PluginExt,
-  PublicPath, Resolve, RspackFuture, RuleSetCondition, RuleSetLogicalConditions, SideEffectOption,
+  Optimization, OutputOptions, ParseOption, ParserOptions, ParserOptionsMap, PathInfo, PublicPath,
+  Resolve, RspackFuture, RuleSetCondition, RuleSetLogicalConditions, SideEffectOption,
   StatsOptions, TrustedTypes, UsedExportsOption, WasmLoading, WasmLoadingType,
 };
 use rspack_fs::{IntermediateFileSystem, ReadableFileSystem, WritableFileSystem};
@@ -58,10 +61,15 @@ use rspack_hash::{HashDigest, HashFunction, HashSalt};
 use rspack_paths::{AssertUtf8, Utf8PathBuf};
 use rspack_regex::RspackRegex;
 use rustc_hash::FxHashMap as HashMap;
+use serde_json::json;
 use target::{get_targets_properties, TargetProperties};
 
+/// Builder trait
 pub trait Builder {
+  /// Target type
   type Item;
+
+  /// Create a builder
   fn builder() -> Self::Item;
 }
 
@@ -73,6 +81,9 @@ impl Builder for Compiler {
   }
 }
 
+/// Builder used to build [`Compiler`].
+///
+/// [`Compiler`]: rspack_core::compiler::Compiler
 #[derive(Default)]
 pub struct CompilerBuilder {
   options_builder: CompilerOptionsBuilder,
@@ -83,6 +94,7 @@ pub struct CompilerBuilder {
 }
 
 impl CompilerBuilder {
+  /// Create a builder with options
   pub fn with_options<V>(options: V) -> Self
   where
     V: Into<CompilerOptionsBuilder>,
@@ -98,16 +110,25 @@ impl CompilerBuilder {
 }
 
 impl CompilerBuilder {
+  /// Set the name of the configuration. Used when loading multiple configurations.
+  ///
+  /// See [`CompilerOptionsBuilder::name`] for more details.
   pub fn name(&mut self, name: String) -> &mut Self {
     self.options_builder.name(name);
     self
   }
 
+  /// Set the target environment.
+  ///
+  /// See [`CompilerOptionsBuilder::target`] for more details.
   pub fn target(&mut self, targets: Targets) -> &mut Self {
     self.options_builder.target(targets);
     self
   }
 
+  /// Set the entry point of the application.
+  ///
+  /// See [`CompilerOptionsBuilder::entry`] for more details.
   pub fn entry<K, V>(&mut self, entry_name: K, entry_description: V) -> &mut Self
   where
     K: Into<String>,
@@ -117,21 +138,33 @@ impl CompilerBuilder {
     self
   }
 
+  /// Set the external libraries that should not be bundled.
+  ///
+  /// See [`CompilerOptionsBuilder::externals`] for more details.
   pub fn externals(&mut self, externals: ExternalItem) -> &mut Self {
     self.options_builder.externals(externals);
     self
   }
 
+  /// Set the type of externals.
+  ///
+  /// See [`CompilerOptionsBuilder::externals_type`] for more details.
   pub fn externals_type(&mut self, externals_type: ExternalType) -> &mut Self {
     self.options_builder.externals_type(externals_type);
     self
   }
 
+  /// Set the presets for external libraries.
+  ///
+  /// See [`CompilerOptionsBuilder::externals_presets`] for more details.
   pub fn externals_presets(&mut self, externals_presets: ExternalsPresets) -> &mut Self {
     self.options_builder.externals_presets(externals_presets);
     self
   }
 
+  /// Set the context in which the compilation should occur.
+  ///
+  /// See [`CompilerOptionsBuilder::context`] for more details.
   pub fn context<V>(&mut self, context: V) -> &mut Self
   where
     V: Into<Context>,
@@ -140,41 +173,65 @@ impl CompilerBuilder {
     self
   }
 
+  /// Set the cache options.
+  ///
+  /// See [`CompilerOptionsBuilder::cache`] for more details.
   pub fn cache(&mut self, cache: CacheOptions) -> &mut Self {
     self.options_builder.cache(cache);
     self
   }
 
+  /// Set the source map configuration.
+  ///
+  /// See [`CompilerOptionsBuilder::devtool`] for more details.
   pub fn devtool(&mut self, devtool: Devtool) -> &mut Self {
     self.options_builder.devtool(devtool);
     self
   }
 
+  /// Set the mode in which Rspack should operate.
+  ///
+  /// See [`CompilerOptionsBuilder::mode`] for more details.
   pub fn mode(&mut self, mode: Mode) -> &mut Self {
     self.options_builder.mode(mode);
     self
   }
 
+  /// Set the resolve configuration.
+  ///
+  /// See [`CompilerOptionsBuilder::resolve`] for more details.
   pub fn resolve(&mut self, resolve: Resolve) -> &mut Self {
     self.options_builder.resolve(resolve);
     self
   }
 
+  /// Set the resolve loader configuration.
+  ///
+  /// See [`CompilerOptionsBuilder::resolve_loader`] for more details.
   pub fn resolve_loader(&mut self, resolve_loader: Resolve) -> &mut Self {
     self.options_builder.resolve_loader(resolve_loader);
     self
   }
 
+  /// Set whether to fail on the first error.
+  ///
+  /// See [`CompilerOptionsBuilder::bail`] for more details.
   pub fn bail(&mut self, bail: bool) -> &mut Self {
     self.options_builder.bail(bail);
     self
   }
 
+  /// Set whether to enable profiling.
+  ///
+  /// See [`CompilerOptionsBuilder::profile`] for more details.
   pub fn profile(&mut self, profile: bool) -> &mut Self {
     self.options_builder.profile(profile);
     self
   }
 
+  /// Set options for module configuration.
+  ///
+  /// See [`CompilerOptionsBuilder::module`] for more details.
   pub fn module<V>(&mut self, module: V) -> &mut Self
   where
     V: Into<ModuleOptionsBuilder>,
@@ -183,6 +240,9 @@ impl CompilerBuilder {
     self
   }
 
+  /// Set options for output.
+  ///
+  /// See [`CompilerOptionsBuilder::output`] for more details.
   pub fn output<V>(&mut self, output: V) -> &mut Self
   where
     V: Into<OutputOptionsBuilder>,
@@ -191,6 +251,9 @@ impl CompilerBuilder {
     self
   }
 
+  /// Set options for optimization.
+  ///
+  /// See [`CompilerOptionsBuilder::optimization`] for more details.
   pub fn optimization<V>(&mut self, optimization: V) -> &mut Self
   where
     V: Into<OptimizationOptionsBuilder>,
@@ -199,6 +262,9 @@ impl CompilerBuilder {
     self
   }
 
+  /// Set options for Node.js environment.
+  ///
+  /// See [`CompilerOptionsBuilder::node`] for more details.
   pub fn node<V>(&mut self, node: V) -> &mut Self
   where
     V: Into<NodeOptionBuilder>,
@@ -207,16 +273,25 @@ impl CompilerBuilder {
     self
   }
 
+  /// Set options for stats.
+  ///
+  /// See [`CompilerOptionsBuilder::stats`] for more details.
   pub fn stats(&mut self, stats: StatsOptions) -> &mut Self {
     self.options_builder.stats(stats);
     self
   }
 
+  /// Set the value of `require.amd` or `define.amd`.
+  ///
+  /// See [`CompilerOptionsBuilder::amd`] for more details.
   pub fn amd(&mut self, amd: String) -> &mut Self {
     self.options_builder.amd(amd);
     self
   }
 
+  /// Set options for experiments.
+  ///
+  /// See [`CompilerOptionsBuilder::experiments`] for more details.
   pub fn experiments<V>(&mut self, experiments: V) -> &mut Self
   where
     V: Into<ExperimentsBuilder>,
@@ -225,21 +300,25 @@ impl CompilerBuilder {
     self
   }
 
+  /// Add a plugin to the compiler.
   pub fn plugin(&mut self, plugin: BoxPlugin) -> &mut Self {
     self.plugins.push(plugin);
     self
   }
 
+  /// Add plugins to the compiler.
   pub fn plugins(&mut self, plugins: impl IntoIterator<Item = BoxPlugin>) -> &mut Self {
     self.plugins.extend(plugins);
     self
   }
 
+  /// Set the input filesystem.
   pub fn input_filesystem(&mut self, input_filesystem: Arc<dyn ReadableFileSystem>) -> &mut Self {
     self.input_filesystem = Some(input_filesystem);
     self
   }
 
+  /// Set the intermediate filesystem.
   pub fn intermediate_filesystem(
     &mut self,
     intermediate_filesystem: Arc<dyn IntermediateFileSystem>,
@@ -248,11 +327,13 @@ impl CompilerBuilder {
     self
   }
 
+  /// Set the output filesystem.
   pub fn output_filesystem(&mut self, output_filesystem: Arc<dyn WritableFileSystem>) -> &mut Self {
     self.output_filesystem = Some(output_filesystem);
     self
   }
 
+  /// Build [`Compiler`] from options and plugins.
   pub fn build(&mut self) -> Compiler {
     let mut builder_context = BuilderContext::default();
     let compiler_options = self.options_builder.build(&mut builder_context);
@@ -325,331 +406,50 @@ impl Builder for Experiments {
   }
 }
 
-/// Options of builtin plugins
-///
-/// The order of this list is strictly ordered with respect to `rspackOptionsApply`.
-#[allow(clippy::enum_variant_names)]
-#[derive(Debug, EnumTag)]
-#[repr(u8)]
-pub(crate) enum BuiltinPluginOptions {
-  // External handling plugins
-  ExternalsPlugin((ExternalType, Vec<ExternalItem>)),
-  NodeTargetPlugin,
-  ElectronTargetPlugin(rspack_plugin_externals::ElectronTargetContext),
-  HttpExternalsRspackPlugin((bool /* css */, bool /* web_async */)),
-
-  // Chunk format and loading plugins
-  ChunkPrefetchPreloadPlugin,
-  CommonJsChunkFormatPlugin,
-  ArrayPushCallbackChunkFormatPlugin,
-  ModuleChunkFormatPlugin,
-  EnableChunkLoadingPlugin(ChunkLoadingType),
-  EnableWasmLoadingPlugin(WasmLoadingType),
-
-  // Runtime and error handling
-  RuntimeChunkPlugin(rspack_plugin_runtime_chunk::RuntimeChunkOptions),
-  NoEmitOnErrorsPlugin,
-
-  // DevTool plugins
-  SourceMapDevToolPlugin(rspack_plugin_devtool::SourceMapDevToolPluginOptions),
-  EvalSourceMapDevToolPlugin(rspack_plugin_devtool::SourceMapDevToolPluginOptions),
-  EvalDevToolModulePlugin(rspack_plugin_devtool::EvalDevToolModulePluginOptions),
-
-  // Core module plugins
-  JavascriptModulesPlugin,
-  JsonModulesPlugin,
-  AssetModulesPlugin,
-  AsyncWebAssemblyModulesPlugin,
-  CssModulesPlugin,
-
-  // Entry and runtime plugins
-  EntryPlugin((String /* entry request */, EntryOptions)),
-  RuntimePlugin,
-  // TODO: add bundler info plugin
-  // BundlerInfoRspackPlugin,
-
-  // Core functionality plugins
-  InferAsyncModulesPlugin,
-  APIPlugin,
-  DataUriPlugin,
-  FileUriPlugin,
-
-  // Optimization plugins
-  EnsureChunkConditionsPlugin,
-  MergeDuplicateChunksPlugin,
-  SideEffectsFlagPlugin,
-  FlagDependencyExportsPlugin,
-  FlagDependencyUsagePlugin(bool),
-  ModuleConcatenationPlugin,
-  MangleExportsPlugin(bool),
-
-  // Experiments
-  // TODO: support lazy compilation
-  // LazyCompilationPlugin,
-
-  // Output plugins
-  EnableLibraryPlugin(LibraryType),
-  // TODO: support split chunks
-  // SplitChunksPlugin,
-  RemoveEmptyChunksPlugin,
-  RealContentHashPlugin,
-
-  // Module and chunk ID plugins
-  NamedModuleIdsPlugin,
-  NaturalModuleIdsPlugin,
-  DeterministicModuleIdsPlugin,
-  NaturalChunkIdsPlugin,
-  NamedChunkIdsPlugin,
-  DeterministicChunkIdsPlugin,
-  OccurrenceChunkIdsPlugin(rspack_ids::OccurrenceChunkIdsPluginOptions),
-
-  // Define and optimization plugins
-  DefinePlugin(rspack_plugin_javascript::define_plugin::DefineValue),
-  AnyMinimizerRspackPlugin(BoxPlugin),
-
-  // TODO: support performance
-  // SizeLimitsPlugin,
-
-  // Cache plugins
-  // MemoryCachePlugin,
-
-  // Worker plugins
-  WorkerPlugin,
-}
-
-#[derive(Default, Debug)]
-pub struct BuilderContext {
-  plugins: Vec<BuiltinPluginOptions>,
-}
-
-impl BuilderContext {
-  pub fn take_plugins(&mut self, compiler_options: &CompilerOptions) -> Vec<BoxPlugin> {
-    self.plugins.sort_by_key(|p| p.tag());
-    let mut plugins = Vec::new();
-    self.plugins.drain(..).for_each(|plugin| match plugin {
-      // External handling plugins
-      BuiltinPluginOptions::ExternalsPlugin((external_type, externals)) => {
-        plugins
-          .push(rspack_plugin_externals::ExternalsPlugin::new(external_type, externals).boxed());
-      }
-      BuiltinPluginOptions::NodeTargetPlugin => {
-        plugins.push(rspack_plugin_externals::node_target_plugin())
-      }
-      BuiltinPluginOptions::ElectronTargetPlugin(context) => {
-        rspack_plugin_externals::electron_target_plugin(context, &mut plugins)
-      }
-      BuiltinPluginOptions::HttpExternalsRspackPlugin((css, web_async)) => {
-        plugins.push(rspack_plugin_externals::http_externals_rspack_plugin(
-          css, web_async,
-        ));
-      }
-
-      // Chunk format and loading plugins
-      BuiltinPluginOptions::ChunkPrefetchPreloadPlugin => {
-        plugins.push(rspack_plugin_runtime::ChunkPrefetchPreloadPlugin::default().boxed());
-      }
-      BuiltinPluginOptions::CommonJsChunkFormatPlugin => {
-        plugins.push(rspack_plugin_runtime::CommonJsChunkFormatPlugin::default().boxed());
-      }
-      BuiltinPluginOptions::ArrayPushCallbackChunkFormatPlugin => {
-        plugins.push(rspack_plugin_runtime::ArrayPushCallbackChunkFormatPlugin::default().boxed());
-      }
-      BuiltinPluginOptions::ModuleChunkFormatPlugin => {
-        plugins.push(rspack_plugin_runtime::ModuleChunkFormatPlugin::default().boxed());
-      }
-      BuiltinPluginOptions::EnableChunkLoadingPlugin(chunk_loading_type) => {
-        rspack_plugin_runtime::enable_chunk_loading_plugin(chunk_loading_type, &mut plugins);
-      }
-      BuiltinPluginOptions::EnableWasmLoadingPlugin(wasm_loading_type) => {
-        plugins.push(rspack_plugin_wasm::enable_wasm_loading_plugin(
-          wasm_loading_type,
-        ));
-      }
-
-      // Runtime and error handling plugins
-      BuiltinPluginOptions::RuntimeChunkPlugin(options) => {
-        plugins.push(rspack_plugin_runtime_chunk::RuntimeChunkPlugin::new(options).boxed());
-      }
-      BuiltinPluginOptions::NoEmitOnErrorsPlugin => {
-        plugins.push(rspack_plugin_no_emit_on_errors::NoEmitOnErrorsPlugin::default().boxed());
-      }
-
-      // DevTool plugins
-      BuiltinPluginOptions::SourceMapDevToolPlugin(options) => {
-        plugins.push(rspack_plugin_devtool::SourceMapDevToolPlugin::new(options).boxed());
-      }
-      BuiltinPluginOptions::EvalSourceMapDevToolPlugin(options) => {
-        plugins.push(rspack_plugin_devtool::EvalSourceMapDevToolPlugin::new(options).boxed());
-      }
-      BuiltinPluginOptions::EvalDevToolModulePlugin(options) => {
-        plugins.push(rspack_plugin_devtool::EvalDevToolModulePlugin::new(options).boxed());
-      }
-
-      // Core module plugins
-      BuiltinPluginOptions::JavascriptModulesPlugin => {
-        plugins.push(rspack_plugin_javascript::JsPlugin::default().boxed());
-      }
-      BuiltinPluginOptions::JsonModulesPlugin => {
-        plugins.push(rspack_plugin_json::JsonPlugin.boxed());
-      }
-      BuiltinPluginOptions::AssetModulesPlugin => {
-        plugins.push(rspack_plugin_asset::AssetPlugin::default().boxed());
-      }
-      BuiltinPluginOptions::AsyncWebAssemblyModulesPlugin => {
-        plugins.push(rspack_plugin_wasm::AsyncWasmPlugin::default().boxed());
-      }
-      BuiltinPluginOptions::CssModulesPlugin => {
-        plugins.push(rspack_plugin_css::CssPlugin::default().boxed());
-      }
-
-      // Entry and runtime plugins
-      BuiltinPluginOptions::EntryPlugin((entry_request, options)) => {
-        plugins.push(
-          rspack_plugin_entry::EntryPlugin::new(
-            compiler_options.context.clone(),
-            entry_request,
-            options,
-          )
-          .boxed(),
-        );
-      }
-      BuiltinPluginOptions::RuntimePlugin => {
-        plugins.push(rspack_plugin_runtime::RuntimePlugin::default().boxed())
-      }
-      // TODO: add bundler info plugin
-      // BuiltinPluginOptions::BundlerInfoRspackPlugin => {}
-
-      // Core functionality plugins
-      BuiltinPluginOptions::InferAsyncModulesPlugin => {
-        plugins.push(rspack_plugin_javascript::InferAsyncModulesPlugin::default().boxed())
-      }
-      BuiltinPluginOptions::APIPlugin => {
-        plugins.push(rspack_plugin_javascript::api_plugin::APIPlugin::default().boxed())
-      }
-      BuiltinPluginOptions::DataUriPlugin => {
-        plugins.push(rspack_plugin_schemes::DataUriPlugin::default().boxed());
-      }
-      BuiltinPluginOptions::FileUriPlugin => {
-        plugins.push(rspack_plugin_schemes::FileUriPlugin::default().boxed());
-      }
-
-      // Optimization plugins
-      BuiltinPluginOptions::EnsureChunkConditionsPlugin => {
-        plugins.push(
-          rspack_plugin_ensure_chunk_conditions::EnsureChunkConditionsPlugin::default().boxed(),
-        );
-      }
-      BuiltinPluginOptions::MergeDuplicateChunksPlugin => {
-        plugins.push(
-          rspack_plugin_merge_duplicate_chunks::MergeDuplicateChunksPlugin::default().boxed(),
-        );
-      }
-      BuiltinPluginOptions::SideEffectsFlagPlugin => {
-        plugins.push(rspack_plugin_javascript::SideEffectsFlagPlugin::default().boxed());
-      }
-      BuiltinPluginOptions::FlagDependencyExportsPlugin => {
-        plugins.push(rspack_plugin_javascript::FlagDependencyExportsPlugin::default().boxed());
-      }
-      BuiltinPluginOptions::FlagDependencyUsagePlugin(value) => {
-        plugins.push(rspack_plugin_javascript::FlagDependencyUsagePlugin::new(value).boxed())
-      }
-      BuiltinPluginOptions::ModuleConcatenationPlugin => {
-        plugins.push(rspack_plugin_javascript::ModuleConcatenationPlugin::default().boxed());
-      }
-      BuiltinPluginOptions::MangleExportsPlugin(value) => {
-        plugins.push(rspack_plugin_javascript::MangleExportsPlugin::new(value).boxed())
-      }
-
-      // Experiments
-      // TODO: support lazy compilation
-      // BuiltinPluginOptions::LazyCompilationPlugin => {
-      // plugins
-      // .push(rspack_plugin_lazy_compilation::plugin::LazyCompilationPlugin::default().boxed());
-      // }
-
-      // Output plugins
-      BuiltinPluginOptions::EnableLibraryPlugin(library_type) => {
-        rspack_plugin_library::enable_library_plugin(library_type, &mut plugins)
-      }
-      // BuiltinPluginOptions::SplitChunksPlugin => {
-      // plugins.push(rspack_plugin_split_chunks::SplitChunksPlugin::default().boxed())
-      // }
-      BuiltinPluginOptions::RemoveEmptyChunksPlugin => {
-        plugins.push(rspack_plugin_remove_empty_chunks::RemoveEmptyChunksPlugin::default().boxed())
-      }
-      BuiltinPluginOptions::RealContentHashPlugin => {
-        plugins.push(rspack_plugin_real_content_hash::RealContentHashPlugin::default().boxed())
-      }
-
-      // Module and chunk ID plugins
-      BuiltinPluginOptions::NamedModuleIdsPlugin => {
-        plugins.push(rspack_ids::NamedModuleIdsPlugin::default().boxed())
-      }
-      BuiltinPluginOptions::NaturalModuleIdsPlugin => {
-        plugins.push(rspack_ids::NaturalModuleIdsPlugin::default().boxed())
-      }
-      BuiltinPluginOptions::DeterministicModuleIdsPlugin => {
-        plugins.push(rspack_ids::DeterministicModuleIdsPlugin::default().boxed())
-      }
-      BuiltinPluginOptions::NaturalChunkIdsPlugin => {
-        plugins.push(rspack_ids::NaturalChunkIdsPlugin::default().boxed())
-      }
-      BuiltinPluginOptions::NamedChunkIdsPlugin => {
-        plugins.push(rspack_ids::NamedChunkIdsPlugin::new(None, None).boxed())
-      }
-      BuiltinPluginOptions::DeterministicChunkIdsPlugin => {
-        plugins.push(rspack_ids::DeterministicChunkIdsPlugin::default().boxed())
-      }
-      BuiltinPluginOptions::OccurrenceChunkIdsPlugin(options) => {
-        plugins.push(rspack_ids::OccurrenceChunkIdsPlugin::new(options).boxed())
-      }
-
-      // Define and optimization plugins
-      BuiltinPluginOptions::DefinePlugin(values) => {
-        plugins.push(rspack_plugin_javascript::define_plugin::DefinePlugin::new(values).boxed())
-      }
-      BuiltinPluginOptions::AnyMinimizerRspackPlugin(plugin) => plugins.push(plugin),
-
-      // TODO: support performance
-      // BuiltinPluginOptions::SizeLimitsPlugin => {
-      // plugins.push(rspack_plugin_size_limits::SizeLimitsPlugin::default().boxed())
-      // }
-
-      // Cache plugins
-      // BuiltinPluginOptions::MemoryCachePlugin => MemoryCachePlugin::default().boxed(),
-
-      // Worker plugins
-      BuiltinPluginOptions::WorkerPlugin => {
-        plugins.push(rspack_plugin_worker::WorkerPlugin::default().boxed())
-      }
-    });
-    plugins
-  }
-}
-
 /// Builder used to build [`CompilerOptions`]
 #[derive(Debug, Default)]
 pub struct CompilerOptionsBuilder {
+  /// Name of the configuration. Used when loading multiple configurations.
   name: Option<String>,
+  /// The environment in which the code should run.
   target: Option<Targets>,
+  /// The entry point of the application.
   entry: IndexMap<String, EntryDescription>,
+  /// External libraries that should not be bundled.
   externals: Option<Vec<ExternalItem>>,
+  /// The type of externals.
   externals_type: Option<ExternalType>,
+  /// Presets for external libraries.
   externals_presets: Option<ExternalsPresets>,
+  /// The context in which the compilation should occur.
   context: Option<Context>,
+  /// Options for caching.
   cache: Option<CacheOptions>,
+  /// The mode in which Rspack should operate.
   mode: Option<Mode>,
+  /// The type of externals.
   resolve: Option<Resolve>,
+  /// The type of externals.
   resolve_loader: Option<Resolve>,
+  /// The type of externals.
   devtool: Option<Devtool>,
+  /// The type of externals.
   profile: Option<bool>,
+  /// Whether to fail on the first error.
   bail: Option<bool>,
+  /// Performance optimization options.
   experiments: Option<ExperimentsBuilder>,
+  /// Options for module configuration.
   module: Option<ModuleOptionsBuilder>,
+  /// Options for stats.
   stats: Option<StatsOptions>,
+  /// Options for output.
   output: Option<OutputOptionsBuilder>,
+  /// Optimization options.
   optimization: Option<OptimizationOptionsBuilder>,
+  /// Options for Node.js environment.
   node: Option<NodeOptionBuilder>,
+  /// The value of `require.amd` or `define.amd`.
   amd: Option<String>,
 }
 
@@ -682,16 +482,19 @@ impl From<&mut CompilerOptionsBuilder> for CompilerOptionsBuilder {
 }
 
 impl CompilerOptionsBuilder {
+  /// Set the name of the configuration. Used when loading multiple configurations.
   pub fn name(&mut self, name: String) -> &mut Self {
     self.name = Some(name);
     self
   }
 
+  /// Set the environment in which the code should run.
   pub fn target(&mut self, targets: Targets) -> &mut Self {
     self.target = Some(targets);
     self
   }
 
+  /// Add an entry point to the configuration.
   pub fn entry<K, V>(&mut self, entry_name: K, entry_description: V) -> &mut Self
   where
     K: Into<String>,
@@ -703,6 +506,7 @@ impl CompilerOptionsBuilder {
     self
   }
 
+  /// Set external libraries that should not be bundled.
   pub fn externals(&mut self, externals: ExternalItem) -> &mut Self {
     match &mut self.externals {
       Some(e) => e.push(externals),
@@ -711,16 +515,19 @@ impl CompilerOptionsBuilder {
     self
   }
 
+  /// Set the type of externals.
   pub fn externals_type(&mut self, externals_type: ExternalType) -> &mut Self {
     self.externals_type = Some(externals_type);
     self
   }
 
+  /// Set presets for external libraries.
   pub fn externals_presets(&mut self, externals_presets: ExternalsPresets) -> &mut Self {
     self.externals_presets = Some(externals_presets);
     self
   }
 
+  /// Set the context in which the compilation should occur.
   pub fn context<V>(&mut self, context: V) -> &mut Self
   where
     V: Into<Context>,
@@ -729,41 +536,49 @@ impl CompilerOptionsBuilder {
     self
   }
 
+  /// Set options for caching.
   pub fn cache(&mut self, cache: CacheOptions) -> &mut Self {
     self.cache = Some(cache);
     self
   }
 
+  /// Set the source map configuration.
   pub fn devtool(&mut self, devtool: Devtool) -> &mut Self {
     self.devtool = Some(devtool);
     self
   }
 
+  /// Set the mode in which Rspack should operate.
   pub fn mode(&mut self, mode: Mode) -> &mut Self {
     self.mode = Some(mode);
     self
   }
 
+  /// Set the resolve configuration.
   pub fn resolve(&mut self, resolve: Resolve) -> &mut Self {
     self.resolve = Some(resolve);
     self
   }
 
+  /// Set the resolve loader configuration.
   pub fn resolve_loader(&mut self, resolve_loader: Resolve) -> &mut Self {
     self.resolve_loader = Some(resolve_loader);
     self
   }
 
+  /// Set whether to fail on the first error.
   pub fn bail(&mut self, bail: bool) -> &mut Self {
     self.bail = Some(bail);
     self
   }
 
+  /// Set whether to enable profiling.
   pub fn profile(&mut self, profile: bool) -> &mut Self {
     self.profile = Some(profile);
     self
   }
 
+  /// Set options for module configuration.
   pub fn module<V>(&mut self, module: V) -> &mut Self
   where
     V: Into<ModuleOptionsBuilder>,
@@ -772,11 +587,13 @@ impl CompilerOptionsBuilder {
     self
   }
 
+  /// Set options for stats.
   pub fn stats(&mut self, stats: StatsOptions) -> &mut Self {
     self.stats = Some(stats);
     self
   }
 
+  /// Set options for output.
   pub fn output<V>(&mut self, output: V) -> &mut Self
   where
     V: Into<OutputOptionsBuilder>,
@@ -785,6 +602,7 @@ impl CompilerOptionsBuilder {
     self
   }
 
+  /// Set options for optimization.
   pub fn optimization<V>(&mut self, optimization: V) -> &mut Self
   where
     V: Into<OptimizationOptionsBuilder>,
@@ -793,6 +611,7 @@ impl CompilerOptionsBuilder {
     self
   }
 
+  /// Set options for Node.js environment.
   pub fn node<V>(&mut self, node: V) -> &mut Self
   where
     V: Into<NodeOptionBuilder>,
@@ -801,11 +620,13 @@ impl CompilerOptionsBuilder {
     self
   }
 
+  /// Set the value of `require.amd` or `define.amd`.
   pub fn amd(&mut self, amd: String) -> &mut Self {
     self.amd = Some(amd);
     self
   }
 
+  /// Set options for experiments.
   pub fn experiments<V>(&mut self, experiments: V) -> &mut Self
   where
     V: Into<ExperimentsBuilder>,
@@ -814,6 +635,15 @@ impl CompilerOptionsBuilder {
     self
   }
 
+  /// Build the options for the compiler, return [`CompilerOptions`].
+  ///
+  /// The returned [`CompilerOptions`] is ready to use for creating compiler.
+  /// Plugins created by this method will be added to the [`BuilderContext`], which will be used to create the compiler.
+  ///
+  /// To create the compiler, you can use [`Compiler::builder`].
+  ///
+  /// [`BuilderContext`]: crate::builder::BuilderContext
+  /// [`CompilerOptions`]: rspack_core::options::CompilerOptions
   pub fn build(&mut self, builder_context: &mut BuilderContext) -> CompilerOptions {
     let name = self.name.take();
     let context = f!(self.context.take(), || {
@@ -1088,28 +918,30 @@ impl CompilerOptionsBuilder {
     };
 
     // apply entry plugin
-    self.entry.drain(..).for_each(|(name, desc)| {
-      let entry_options = EntryOptions {
-        name: Some(name),
-        runtime: desc.runtime.map(EntryRuntime::String),
-        chunk_loading: desc.chunk_loading,
-        async_chunks: desc.async_chunks,
-        public_path: desc.public_path,
-        base_uri: desc.base_uri,
-        filename: desc.filename,
-        library: None,
-        depend_on: desc.depend_on,
-        layer: None,
-      };
-      desc.import.into_iter().for_each(|import| {
-        builder_context
-          .plugins
-          .push(BuiltinPluginOptions::EntryPlugin((
-            import,
-            entry_options.clone(),
-          )));
+    std::mem::take(&mut self.entry)
+      .into_iter()
+      .for_each(|(name, desc)| {
+        let entry_options = EntryOptions {
+          name: Some(name),
+          runtime: desc.runtime.map(EntryRuntime::String),
+          chunk_loading: desc.chunk_loading,
+          async_chunks: desc.async_chunks,
+          public_path: desc.public_path,
+          base_uri: desc.base_uri,
+          filename: desc.filename,
+          library: None,
+          depend_on: desc.depend_on,
+          layer: None,
+        };
+        desc.import.into_iter().for_each(|import| {
+          builder_context
+            .plugins
+            .push(BuiltinPluginOptions::EntryPlugin((
+              import,
+              entry_options.clone(),
+            )));
+        });
       });
-    });
 
     builder_context
       .plugins
@@ -1324,14 +1156,35 @@ fn get_resolve_loader_defaults() -> Resolve {
   }
 }
 
+/// Builder used to build [`NodeOption`].
+///
+/// [`NodeOption`]: rspack_core::options::NodeOption
 #[derive(Debug)]
 pub enum NodeOptionBuilder {
+  /// Set options for Node.js environment.
   True {
+    /// Set the `__dirname` for Node.js environment.
     dirname: Option<NodeDirnameOption>,
+    /// Set the `global` for Node.js environment.
     global: Option<NodeGlobalOption>,
+    /// Set the `__filename` for Node.js environment.
     filename: Option<NodeFilenameOption>,
   },
+  /// Disable Node.js environment.
   False,
+}
+
+impl From<Option<NodeOption>> for NodeOptionBuilder {
+  fn from(value: Option<NodeOption>) -> Self {
+    match value {
+      Some(node_option) => NodeOptionBuilder::True {
+        dirname: Some(node_option.dirname),
+        global: Some(node_option.global),
+        filename: Some(node_option.filename),
+      },
+      None => NodeOptionBuilder::False,
+    }
+  }
 }
 
 impl From<&mut NodeOptionBuilder> for NodeOptionBuilder {
@@ -1362,11 +1215,13 @@ impl Default for NodeOptionBuilder {
 }
 
 impl NodeOptionBuilder {
+  /// Disable Node.js environment.
   pub fn disabled(&mut self) -> &mut Self {
     *self = NodeOptionBuilder::False;
     self
   }
 
+  /// Set the `__dirname` for Node.js environment.
   pub fn dirname(&mut self, dirname: NodeDirnameOption) -> &mut Self {
     match self {
       NodeOptionBuilder::True { dirname: d, .. } => {
@@ -1383,6 +1238,7 @@ impl NodeOptionBuilder {
     self
   }
 
+  /// Set the `global` for Node.js environment.
   pub fn global(&mut self, global: NodeGlobalOption) -> &mut Self {
     match self {
       NodeOptionBuilder::True { global: g, .. } => {
@@ -1399,6 +1255,7 @@ impl NodeOptionBuilder {
     self
   }
 
+  /// Set the `__filename` for Node.js environment.
   pub fn filename(&mut self, filename: NodeFilenameOption) -> &mut Self {
     match self {
       NodeOptionBuilder::True { filename: f, .. } => {
@@ -1415,6 +1272,9 @@ impl NodeOptionBuilder {
     self
   }
 
+  /// Build [`NodeOption`] from options.
+  ///
+  /// [`NodeOption`]: rspack_core::options::NodeOption
   pub fn build(
     &mut self,
     target_properties: &TargetProperties,
@@ -1462,19 +1322,36 @@ impl NodeOptionBuilder {
   }
 }
 
-/// Builder used to build [`ModuleOptions`]
+/// Builder used to build [`ModuleOptions`].
+///
+/// [`ModuleOptions`]: rspack_core::options::ModuleOptions
 #[derive(Debug, Default)]
 pub struct ModuleOptionsBuilder {
+  /// An array of rules that match the module's requests when it is created.
   rules: Vec<ModuleRule>,
+  /// Configure all parsers' options in one place with module.parser.
   parser: Option<ParserOptionsMap>,
+  /// Configure all generators' options in one place with module.generator.
   generator: Option<GeneratorOptionsMap>,
+  /// Keep module mechanism of the matched modules as-is, such as module.exports, require, import.
   no_parse: Option<ModuleNoParseRules>,
+}
+
+impl From<ModuleOptions> for ModuleOptionsBuilder {
+  fn from(value: ModuleOptions) -> Self {
+    ModuleOptionsBuilder {
+      rules: value.rules,
+      parser: value.parser,
+      generator: value.generator,
+      no_parse: value.no_parse,
+    }
+  }
 }
 
 impl From<&mut ModuleOptionsBuilder> for ModuleOptionsBuilder {
   fn from(value: &mut ModuleOptionsBuilder) -> Self {
     ModuleOptionsBuilder {
-      rules: value.rules.drain(..).collect(),
+      rules: std::mem::take(&mut value.rules),
       parser: value.parser.take(),
       generator: value.generator.take(),
       no_parse: value.no_parse.take(),
@@ -1483,16 +1360,19 @@ impl From<&mut ModuleOptionsBuilder> for ModuleOptionsBuilder {
 }
 
 impl ModuleOptionsBuilder {
+  /// Add a rule to the module.
   pub fn rule(&mut self, rule: ModuleRule) -> &mut Self {
     self.rules.push(rule);
     self
   }
 
+  /// Add multiple rules to the module.
   pub fn rules(&mut self, mut rules: Vec<ModuleRule>) -> &mut Self {
     self.rules.append(&mut rules);
     self
   }
 
+  /// Set the parser options for the module.
   pub fn parser(&mut self, parser: ParserOptionsMap) -> &mut Self {
     match &mut self.parser {
       Some(p) => p.extend(parser.clone()),
@@ -1501,6 +1381,7 @@ impl ModuleOptionsBuilder {
     self
   }
 
+  /// Set the generator options for the module.
   pub fn generator(&mut self, generator: GeneratorOptionsMap) -> &mut Self {
     match &mut self.generator {
       Some(g) => g.extend(generator.clone()),
@@ -1509,11 +1390,15 @@ impl ModuleOptionsBuilder {
     self
   }
 
+  /// Set the no_parse options for the module.
   pub fn no_parse(&mut self, no_parse: ModuleNoParseRules) -> &mut Self {
     self.no_parse = Some(no_parse);
     self
   }
 
+  /// Build [`ModuleOptions`] from options.
+  ///
+  /// [`ModuleOptions`]: rspack_core::options::ModuleOptions
   pub fn build(
     &mut self,
     _builder_context: &mut BuilderContext,
@@ -1914,56 +1799,159 @@ fn default_rules(async_web_assembly: bool, css: bool) -> Vec<ModuleRule> {
   rules
 }
 
-/// Builder used to build [`OutputOptions`]
+/// Builder used to build [`OutputOptions`].
+///
+/// [`OutputOptions`]: rspack_core::options::OutputOptions
 #[derive(Debug, Default)]
 pub struct OutputOptionsBuilder {
+  /// Set the output path.
   path: Option<Utf8PathBuf>,
+  /// Set the pathinfo option.
   pathinfo: Option<PathInfo>,
+  /// Set the clean option.
   clean: Option<CleanOptions>,
+  /// Set the public path.
   public_path: Option<PublicPath>,
+  /// Set the asset module filename.
   asset_module_filename: Option<Filename>,
+  /// Set the wasm loading.
   wasm_loading: Option<WasmLoading>,
+  /// Set the wasm module filename.
   webassembly_module_filename: Option<FilenameTemplate>,
+  /// Set the unique name.
   unique_name: Option<String>,
+  /// Set the chunk loading.
   chunk_loading: Option<ChunkLoading>,
+  /// Set the chunk loading global.
   chunk_loading_global: Option<String>,
+  /// Set the chunk load timeout.
   chunk_load_timeout: Option<u32>,
+  /// Set the chunk format.
   chunk_format: Option<String>,
+  /// Set the charset.
   charset: Option<bool>,
+  /// Set the filename.
   filename: Option<Filename>,
+  /// Set the chunk filename.
   chunk_filename: Option<Filename>,
+  /// Set the cross origin loading.
   cross_origin_loading: Option<CrossOriginLoading>,
+  /// Set the css filename.
   css_filename: Option<Filename>,
+  /// Set the css chunk filename.
   css_chunk_filename: Option<Filename>,
+  /// Set the hot update main filename.
   hot_update_main_filename: Option<FilenameTemplate>,
+  /// Set the hot update chunk filename.
   hot_update_chunk_filename: Option<FilenameTemplate>,
+  /// Set the hot update global.
   hot_update_global: Option<String>,
+  /// Set the library.
   library: Option<LibraryOptions>,
+  /// Set the enabled library types.
   enabled_library_types: Option<Vec<LibraryType>>,
+  /// Set the enabled chunk loading types.
   enabled_chunk_loading_types: Option<Vec<ChunkLoadingType>>,
+  /// Set the enabled wasm loading types.
   enabled_wasm_loading_types: Option<Vec<WasmLoadingType>>,
+  /// Set the strict module error handling.
   strict_module_error_handling: Option<bool>,
+  /// Set the global object.
   global_object: Option<String>,
+  /// Set the import function name.
   import_function_name: Option<String>,
+  /// Set the import meta name.
   import_meta_name: Option<String>,
+  /// Set the iife.
   iife: Option<bool>,
+  /// Set the module.
   module: Option<bool>,
+  /// Set the trusted types.
   trusted_types: Option<TrustedTypes>,
+  /// Set the source map filename.
   source_map_filename: Option<FilenameTemplate>,
+  /// Set the hash function.
   hash_function: Option<HashFunction>,
+  /// Set the hash digest.
   hash_digest: Option<HashDigest>,
+  /// Set the hash digest length.
   hash_digest_length: Option<usize>,
+  /// Set the hash salt.
   hash_salt: Option<HashSalt>,
+  /// Set the async chunks.
   async_chunks: Option<bool>,
+  /// Set the worker chunk loading.
   worker_chunk_loading: Option<ChunkLoading>,
+  /// Set the worker wasm loading.
   worker_wasm_loading: Option<WasmLoading>,
+  /// Set the worker public path.
   worker_public_path: Option<String>,
+  /// Set the script type.
   script_type: Option<String>,
+  /// Set the devtool namespace.
   devtool_namespace: Option<String>,
+  /// Set the devtool module filename template.
   devtool_module_filename_template: Option<FilenameTemplate>,
+  /// Set the devtool fallback module filename template.
   devtool_fallback_module_filename_template: Option<FilenameTemplate>,
+  /// Set the environment.
   environment: Option<Environment>,
+  /// Set the compare before emit.
   compare_before_emit: Option<bool>,
+}
+
+impl From<OutputOptions> for OutputOptionsBuilder {
+  fn from(value: OutputOptions) -> Self {
+    OutputOptionsBuilder {
+      path: Some(value.path),
+      pathinfo: Some(value.pathinfo),
+      clean: Some(value.clean),
+      public_path: Some(value.public_path),
+      asset_module_filename: Some(value.asset_module_filename),
+      wasm_loading: Some(value.wasm_loading),
+      webassembly_module_filename: Some(value.webassembly_module_filename),
+      unique_name: Some(value.unique_name),
+      chunk_loading: Some(value.chunk_loading),
+      chunk_loading_global: Some(value.chunk_loading_global),
+      chunk_load_timeout: Some(value.chunk_load_timeout),
+      chunk_format: None,
+      charset: Some(value.charset),
+      filename: Some(value.filename),
+      chunk_filename: Some(value.chunk_filename),
+      cross_origin_loading: Some(value.cross_origin_loading),
+      css_filename: Some(value.css_filename),
+      css_chunk_filename: Some(value.css_chunk_filename),
+      hot_update_main_filename: Some(value.hot_update_main_filename),
+      hot_update_chunk_filename: Some(value.hot_update_chunk_filename),
+      hot_update_global: Some(value.hot_update_global),
+      library: value.library,
+      enabled_library_types: value.enabled_library_types,
+      strict_module_error_handling: Some(value.strict_module_error_handling),
+      global_object: Some(value.global_object),
+      import_function_name: Some(value.import_function_name),
+      import_meta_name: Some(value.import_meta_name),
+      iife: Some(value.iife),
+      module: Some(value.module),
+      trusted_types: value.trusted_types,
+      source_map_filename: Some(value.source_map_filename),
+      hash_function: Some(value.hash_function),
+      hash_digest: Some(value.hash_digest),
+      hash_digest_length: Some(value.hash_digest_length),
+      hash_salt: Some(value.hash_salt),
+      async_chunks: Some(value.async_chunks),
+      worker_chunk_loading: Some(value.worker_chunk_loading),
+      worker_wasm_loading: Some(value.worker_wasm_loading),
+      worker_public_path: Some(value.worker_public_path),
+      script_type: Some(value.script_type),
+      devtool_namespace: None,
+      devtool_module_filename_template: None,
+      devtool_fallback_module_filename_template: None,
+      environment: Some(value.environment),
+      compare_before_emit: Some(value.compare_before_emit),
+      enabled_chunk_loading_types: None,
+      enabled_wasm_loading_types: None,
+    }
+  }
 }
 
 impl From<&mut OutputOptionsBuilder> for OutputOptionsBuilder {
@@ -2023,6 +2011,9 @@ impl From<&mut OutputOptionsBuilder> for OutputOptionsBuilder {
 }
 
 impl OutputOptionsBuilder {
+  /// The output directory as an absolute path.
+  ///
+  /// Default set to `std::env::current_dir().unwrap().join("dist")`.
   pub fn path<V>(&mut self, path: V) -> &mut Self
   where
     V: Into<Utf8PathBuf>,
@@ -2031,221 +2022,289 @@ impl OutputOptionsBuilder {
     self
   }
 
+  /// Tells Rspack to include comments in bundles with information about the contained modules.
+  ///
+  /// Default set to `PathInfo::Bool(true)`.
   pub fn pathinfo(&mut self, pathinfo: PathInfo) -> &mut Self {
     self.pathinfo = Some(pathinfo);
     self
   }
 
+  /// Before generating the products, whether delete all files in the output directory.
+  ///
+  /// Default set to `CleanOptions::CleanAll(false)`.
   pub fn clean(&mut self, clean: CleanOptions) -> &mut Self {
     self.clean = Some(clean);
     self
   }
 
+  /// This option determines the URL prefix of the referenced resource, such as: image, file, etc.
   pub fn public_path(&mut self, public_path: PublicPath) -> &mut Self {
     self.public_path = Some(public_path);
     self
   }
 
+  /// This option determines the name of each output bundle.
+  ///
+  /// Default set to `"[hash][ext][query]"`.
   pub fn asset_module_filename(&mut self, filename: Filename) -> &mut Self {
     self.asset_module_filename = Some(filename);
     self
   }
 
+  /// This option determines the name of each output wasm bundle.
+  ///
+  /// Default set to [`WasmLoadingType::Fetch`].
+  ///
+  /// [`WasmLoadingType`]: rspack_core::options::WasmLoadingType
   pub fn wasm_loading(&mut self, loading: WasmLoading) -> &mut Self {
     self.wasm_loading = Some(loading);
     self
   }
 
+  /// This option determines the name of each output wasm bundle.
+  ///
+  /// Default set to `"[hash].module.wasm"`.
   pub fn webassembly_module_filename(&mut self, filename: FilenameTemplate) -> &mut Self {
     self.webassembly_module_filename = Some(filename);
     self
   }
 
+  /// A unique name of the Rspack build to avoid multiple Rspack runtimes to conflict when using globals.
   pub fn unique_name(&mut self, name: String) -> &mut Self {
     self.unique_name = Some(name);
     self
   }
 
+  /// This method to load chunks (methods included by default are 'jsonp' (web), 'import' (ESM), 'importScripts' (webworker), 'require' (sync node.js), 'async-node' (async node.js).
   pub fn chunk_loading(&mut self, loading: ChunkLoading) -> &mut Self {
     self.chunk_loading = Some(loading);
     self
   }
 
+  /// The global variable is used by Rspack for loading chunks.
+  ///
+  /// Determined by output.uniqueName default.
   pub fn chunk_loading_global(&mut self, global: String) -> &mut Self {
     self.chunk_loading_global = Some(global);
     self
   }
 
+  /// The Number of milliseconds before chunk request timed out.
+  ///
+  /// Default set to `120000`.
   pub fn chunk_load_timeout(&mut self, timeout: u32) -> &mut Self {
     self.chunk_load_timeout = Some(timeout);
     self
   }
 
+  /// The format of chunks (formats included by default are 'array-push' (web/webworker), 'commonjs' (node.js), 'module' (ESM).
   pub fn chunk_format(&mut self, chunk_format: String) -> &mut Self {
     self.chunk_format = Some(chunk_format);
     self
   }
 
+  /// Add charset="utf-8" to the HTML <script> tag.
+  ///
+  /// Default set to `true`.
   pub fn charset(&mut self, charset: bool) -> &mut Self {
     self.charset = Some(charset);
     self
   }
 
+  /// Set the name of each asset resource output bundle.
   pub fn filename(&mut self, filename: Filename) -> &mut Self {
     self.filename = Some(filename);
     self
   }
 
+  /// Set the name of non-initial chunk files.
   pub fn chunk_filename(&mut self, filename: Filename) -> &mut Self {
     self.chunk_filename = Some(filename);
     self
   }
 
+  /// Set the [crossorigin attribute](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/script) for dynamically loaded chunks.
   pub fn cross_origin_loading(&mut self, loading: CrossOriginLoading) -> &mut Self {
     self.cross_origin_loading = Some(loading);
     self
   }
 
+  /// Set the name of CSS output files on disk.
   pub fn css_filename(&mut self, filename: Filename) -> &mut Self {
     self.css_filename = Some(filename);
     self
   }
 
+  /// Set the name of non-initial CSS output files on disk.
   pub fn css_chunk_filename(&mut self, filename: Filename) -> &mut Self {
     self.css_chunk_filename = Some(filename);
     self
   }
 
+  /// Customize the main hot update filename. [fullhash] and [runtime] are available as placeholder.
+  ///
+  /// Default set to `"[runtime].[fullhash].hot-update.json"`.
   pub fn hot_update_main_filename(&mut self, filename: FilenameTemplate) -> &mut Self {
     self.hot_update_main_filename = Some(filename);
     self
   }
 
+  /// Customize the filenames of hot update chunks.
+  ///
+  /// Default set to `"[id].[fullhash].hot-update.js"`.
   pub fn hot_update_chunk_filename(&mut self, filename: FilenameTemplate) -> &mut Self {
     self.hot_update_chunk_filename = Some(filename);
     self
   }
 
+  /// Set the global variable is used by Rspack for loading hot updates.
+  ///
+  /// Determined by output.uniqueName default.
   pub fn hot_update_global(&mut self, global: String) -> &mut Self {
     self.hot_update_global = Some(global);
     self
   }
 
+  /// Set the name of each output bundle.
   pub fn library(&mut self, library: LibraryOptions) -> &mut Self {
     self.library = Some(library);
     self
   }
 
+  /// Set list of library types enabled for use by entry points.
   pub fn enabled_library_types(&mut self, types: Vec<LibraryType>) -> &mut Self {
     self.enabled_library_types = Some(types);
     self
   }
 
+  /// Set list of chunk loading types enabled for use by entry points.
   pub fn enabled_chunk_loading_types(&mut self, types: Vec<ChunkLoadingType>) -> &mut Self {
     self.enabled_chunk_loading_types = Some(types);
     self
   }
 
+  /// Set list of wasm loading types enabled for use by entry points.
   pub fn enabled_wasm_loading_types(&mut self, types: Vec<WasmLoadingType>) -> &mut Self {
     self.enabled_wasm_loading_types = Some(types);
     self
   }
 
+  /// Set whether to enable strict module error handling.
   pub fn strict_module_error_handling(&mut self, strict: bool) -> &mut Self {
     self.strict_module_error_handling = Some(strict);
     self
   }
 
+  /// Set global object that indicates what global object will be used to mount the library.
   pub fn global_object(&mut self, object: String) -> &mut Self {
     self.global_object = Some(object);
     self
   }
 
+  /// Set the name of the import function.
   pub fn import_function_name(&mut self, name: String) -> &mut Self {
     self.import_function_name = Some(name);
     self
   }
 
+  /// Set the name of the import meta.
   pub fn import_meta_name(&mut self, name: String) -> &mut Self {
     self.import_meta_name = Some(name);
     self
   }
 
+  /// Set whether to tell Rspack to add IIFE wrapper around emitted code.
   pub fn iife(&mut self, iife: bool) -> &mut Self {
     self.iife = Some(iife);
     self
   }
 
+  /// Set whether to output JavaScript files as module type.
   pub fn module(&mut self, module: bool) -> &mut Self {
     self.module = Some(module);
     self
   }
 
+  /// Set controls [Trusted Types](https://web.dev/articles/trusted-types) compatibility.
   pub fn trusted_types(&mut self, trusted_types: TrustedTypes) -> &mut Self {
     self.trusted_types = Some(trusted_types);
     self
   }
 
+  /// Set the name of the source map file.
   pub fn source_map_filename(&mut self, filename: FilenameTemplate) -> &mut Self {
     self.source_map_filename = Some(filename);
     self
   }
 
+  /// Set the hash function.
   pub fn hash_function(&mut self, function: HashFunction) -> &mut Self {
     self.hash_function = Some(function);
     self
   }
 
+  /// Set the hash digest.
   pub fn hash_digest(&mut self, digest: HashDigest) -> &mut Self {
     self.hash_digest = Some(digest);
     self
   }
 
+  /// Set the hash digest length.
   pub fn hash_digest_length(&mut self, length: usize) -> &mut Self {
     self.hash_digest_length = Some(length);
     self
   }
 
+  /// Set the hash salt.
   pub fn hash_salt(&mut self, salt: HashSalt) -> &mut Self {
     self.hash_salt = Some(salt);
     self
   }
 
+  /// Set whether to enable async chunks.
   pub fn async_chunks(&mut self, async_chunks: bool) -> &mut Self {
     self.async_chunks = Some(async_chunks);
     self
   }
 
+  /// Set the chunk loading type for worker.
   pub fn worker_chunk_loading(&mut self, loading: ChunkLoading) -> &mut Self {
     self.worker_chunk_loading = Some(loading);
     self
   }
 
+  /// Set the wasm loading type for worker.
   pub fn worker_wasm_loading(&mut self, loading: WasmLoading) -> &mut Self {
     self.worker_wasm_loading = Some(loading);
     self
   }
 
+  /// Set the public path for Worker.
   pub fn worker_public_path(&mut self, path: String) -> &mut Self {
     self.worker_public_path = Some(path);
     self
   }
 
+  /// Set the type of the script.
   pub fn script_type(&mut self, script_type: String) -> &mut Self {
     self.script_type = Some(script_type);
     self
   }
 
+  /// Set the namespace of the devtool.
   pub fn devtool_namespace(&mut self, namespace: String) -> &mut Self {
     self.devtool_namespace = Some(namespace);
     self
   }
 
+  /// Set the template of the devtool module filename.
   pub fn devtool_module_filename_template(&mut self, filename: FilenameTemplate) -> &mut Self {
     self.devtool_module_filename_template = Some(filename);
     self
   }
 
+  /// Set the template of the devtool fallback module filename.
   pub fn devtool_fallback_module_filename_template(
     &mut self,
     filename: FilenameTemplate,
@@ -2253,16 +2312,22 @@ impl OutputOptionsBuilder {
     self.devtool_fallback_module_filename_template = Some(filename);
     self
   }
+
+  /// Set the environment.
   pub fn environment(&mut self, environment: Environment) -> &mut Self {
     self.environment = Some(environment);
     self
   }
 
+  /// Set whether to compare the emitted code before emitting.
   pub fn compare_before_emit(&mut self, compare: bool) -> &mut Self {
     self.compare_before_emit = Some(compare);
     self
   }
 
+  /// Build [`OutputOptions`] from builder.
+  ///
+  /// [`OutputOptions`]: rspack_core::options::OutputOptions
   #[allow(clippy::too_many_arguments, clippy::fn_params_excessive_bools)]
   pub fn build(
     &mut self,
@@ -2745,28 +2810,75 @@ impl OutputOptionsBuilder {
   }
 }
 
-/// Builder used to build options for optimization plugins
+/// Builder used to build options for optimization plugins.
+///
+/// See [`OptimizationOptions`] for more details.
+///
+/// [`OptimizationOptions`]: rspack_core::options::Optimization
 #[derive(Debug, Default)]
 pub struct OptimizationOptionsBuilder {
+  /// Detect and remove modules from chunks these modules are already included in all parents.
   remove_available_modules: Option<bool>,
+  /// Remove empty chunks generated in the compilation.
   remove_empty_chunks: Option<bool>,
+  /// Merge chunks which contain the same modules.
   merge_duplicate_chunks: Option<bool>,
+  /// Which algorithm to use when choosing module ids.
   module_ids: Option<String>,
+  /// Which algorithm to use when choosing chunk ids.
   chunk_ids: Option<String>,
+  /// Whether to enable minimize.
   minimize: Option<bool>,
+  /// Minimizer.
   minimizer: Option<Vec<BuiltinPluginOptions>>,
+  /// Whether to enable side effects.
   side_effects: Option<SideEffectOption>,
+  /// Whether to enable provided exports.
   provided_exports: Option<bool>,
+  /// Whether to enable used exports.
   used_exports: Option<UsedExportsOption>,
+  /// Whether to enable inner graph.
   inner_graph: Option<bool>,
+  /// Whether to enable mangle exports.
   mangle_exports: Option<MangleExportsOption>,
+  /// Whether to enable concatenate modules.
   concatenate_modules: Option<bool>,
+  /// Whether to enable real content hash.
   real_content_hash: Option<bool>,
+  /// Whether to enable avoid entry iife.
   avoid_entry_iife: Option<bool>,
+  /// Node env.
   node_env: Option<String>,
+  /// Whether to emit on errors.
   emit_on_errors: Option<bool>,
+  /// Runtime chunk.
   runtime_chunk: Option<rspack_plugin_runtime_chunk::RuntimeChunkOptions>,
   // TODO: split chunks
+}
+
+impl From<Optimization> for OptimizationOptionsBuilder {
+  fn from(value: Optimization) -> Self {
+    OptimizationOptionsBuilder {
+      remove_available_modules: Some(value.remove_available_modules),
+      side_effects: Some(value.side_effects),
+      provided_exports: Some(value.provided_exports),
+      used_exports: Some(value.used_exports),
+      inner_graph: Some(value.inner_graph),
+      mangle_exports: Some(value.mangle_exports),
+      concatenate_modules: Some(value.concatenate_modules),
+      avoid_entry_iife: Some(value.avoid_entry_iife),
+      remove_empty_chunks: None,
+      merge_duplicate_chunks: None,
+      module_ids: None,
+      chunk_ids: None,
+      minimize: None,
+      minimizer: None,
+      real_content_hash: None,
+      node_env: None,
+      emit_on_errors: None,
+      runtime_chunk: None,
+    }
+  }
 }
 
 impl From<&mut OptimizationOptionsBuilder> for OptimizationOptionsBuilder {
@@ -2795,36 +2907,46 @@ impl From<&mut OptimizationOptionsBuilder> for OptimizationOptionsBuilder {
 }
 
 impl OptimizationOptionsBuilder {
+  /// Set whether to detect and remove modules from chunks these modules are already included in all parents.
   pub fn remove_available_modules(&mut self, value: bool) -> &mut Self {
     self.remove_available_modules = Some(value);
     self
   }
 
+  /// Set whether to remove empty chunks generated in the compilation.
   pub fn remove_empty_chunks(&mut self, value: bool) -> &mut Self {
     self.remove_empty_chunks = Some(value);
     self
   }
 
+  /// Set whether to merge chunks which contain the same modules.
   pub fn merge_duplicate_chunks(&mut self, value: bool) -> &mut Self {
     self.merge_duplicate_chunks = Some(value);
     self
   }
 
+  /// Set which algorithm to use when choosing module ids.
   pub fn module_ids(&mut self, value: String) -> &mut Self {
     self.module_ids = Some(value);
     self
   }
 
+  /// Set which algorithm to use when choosing chunk ids.
   pub fn chunk_ids(&mut self, value: String) -> &mut Self {
     self.chunk_ids = Some(value);
     self
   }
 
+  /// Set whether to enable minimize.
+  ///
+  /// Default set to `true` in production mode.
+  /// Default set to `false` in development mode.
   pub fn minimize(&mut self, value: bool) -> &mut Self {
     self.minimize = Some(value);
     self
   }
 
+  /// Set the minimizer.
   pub fn minimizer(&mut self, value: Vec<BoxPlugin>) -> &mut Self {
     self.minimizer = Some(
       value
@@ -2835,46 +2957,79 @@ impl OptimizationOptionsBuilder {
     self
   }
 
+  /// Set whether to enable side effects.
+  ///
+  /// Default set to `true` in production mode.
+  /// Default set to `false` in development mode.
   pub fn side_effects(&mut self, value: SideEffectOption) -> &mut Self {
     self.side_effects = Some(value);
     self
   }
 
+  /// Set whether to enable provided exports.
+  ///
+  /// After enabling, Rspack will analyze which exports the module provides, including re-exported modules.
+  ///
+  /// Default set to `true`.
   pub fn provided_exports(&mut self, value: bool) -> &mut Self {
     self.provided_exports = Some(value);
     self
   }
 
+  /// Set whether to enable used exports.
+  ///
+  /// Default set to `true` in production mode.
+  /// Default set to `false` in development mode.
   pub fn used_exports(&mut self, value: UsedExportsOption) -> &mut Self {
     self.used_exports = Some(value);
     self
   }
 
+  /// Set whether to enable inner graph.
+  ///
+  /// Default set to `true` in production mode.
+  /// Default set to `false` in development mode.
   pub fn inner_graph(&mut self, value: bool) -> &mut Self {
     self.inner_graph = Some(value);
     self
   }
 
+  /// Set whether to enable mangle exports.
+  ///
+  /// Default set to `deterministic` in production mode.
+  /// Default set to `false` in development mode.
   pub fn mangle_exports(&mut self, value: MangleExportsOption) -> &mut Self {
     self.mangle_exports = Some(value);
     self
   }
 
+  /// Set whether to enable concatenate modules.
+  ///
+  /// Default set to `true` in production mode.
+  /// Default set to `false` in development mode.
   pub fn concatenate_modules(&mut self, value: bool) -> &mut Self {
     self.concatenate_modules = Some(value);
     self
   }
 
+  /// Set whether to enable real content hash.
+  ///
+  /// Default set to `true` in production mode.
+  /// Default set to `false` in development mode.
   pub fn real_content_hash(&mut self, value: bool) -> &mut Self {
     self.real_content_hash = Some(value);
     self
   }
 
+  /// Set whether to enable avoid entry iife.
+  ///
+  /// Default set to `false`.
   pub fn avoid_entry_iife(&mut self, value: bool) -> &mut Self {
     self.avoid_entry_iife = Some(value);
     self
   }
 
+  /// Set the node env.
   pub fn node_env<V>(&mut self, value: V) -> &mut Self
   where
     V: Into<String>,
@@ -2883,11 +3038,16 @@ impl OptimizationOptionsBuilder {
     self
   }
 
+  /// Set whether to emit on errors.
+  ///
+  /// Default set to `true` in production mode.
+  /// Default set to `false` in development mode.
   pub fn emit_on_errors(&mut self, value: bool) -> &mut Self {
     self.emit_on_errors = Some(value);
     self
   }
 
+  /// Set the runtime chunk.
   pub fn runtime_chunk(
     &mut self,
     value: rspack_plugin_runtime_chunk::RuntimeChunkOptions,
@@ -2896,6 +3056,9 @@ impl OptimizationOptionsBuilder {
     self
   }
 
+  /// Build [`Optimization`] from options.
+  ///
+  /// [`Optimization`]: rspack_core::options::Optimization
   pub fn build(
     &mut self,
     builder_context: &mut BuilderContext,
@@ -3124,22 +3287,49 @@ impl OptimizationOptionsBuilder {
   }
 }
 
-/// Builder used to build [`Experiments`]
+/// Builder used to build [`Experiments`].
+///
+/// [`Experiments`]: rspack_core::options::Experiments
 #[derive(Debug, Default)]
 pub struct ExperimentsBuilder {
+  /// Whether to enable module layers feature.  
   layers: Option<bool>,
+  /// Incremental passes.
   incremental: Option<IncrementalPasses>,
+  /// Whether to enable top level await.
   top_level_await: Option<bool>,
+  /// Rspack future.
   rspack_future: Option<RspackFuture>,
+  /// Cache options.
   cache: Option<ExperimentCacheOptions>,
-
-  // Builder specific
+  /// Whether to enable output module.
   output_module: Option<bool>,
+  /// Whether to enable future defaults.
   future_defaults: Option<bool>,
+  /// Whether to enable css.
   css: Option<bool>,
+  /// Whether to enable parallel code splitting.
   parallel_code_splitting: Option<bool>,
+  /// Whether to enable async web assembly.
   async_web_assembly: Option<bool>,
   // TODO: lazy compilation
+}
+
+impl From<Experiments> for ExperimentsBuilder {
+  fn from(value: Experiments) -> Self {
+    ExperimentsBuilder {
+      layers: Some(value.layers),
+      incremental: Some(value.incremental),
+      top_level_await: Some(value.top_level_await),
+      rspack_future: Some(value.rspack_future),
+      cache: Some(value.cache),
+      parallel_code_splitting: Some(value.parallel_code_splitting),
+      output_module: None,
+      future_defaults: None,
+      css: None,
+      async_web_assembly: None,
+    }
+  }
 }
 
 impl From<&mut ExperimentsBuilder> for ExperimentsBuilder {
@@ -3160,46 +3350,57 @@ impl From<&mut ExperimentsBuilder> for ExperimentsBuilder {
 }
 
 impl ExperimentsBuilder {
+  /// Set whether to enable layers.
   pub fn layers(&mut self, layers: bool) -> &mut Self {
     self.layers = Some(layers);
     self
   }
 
+  /// Set the incremental passes.
   pub fn incremental(&mut self, incremental: IncrementalPasses) -> &mut Self {
     self.incremental = Some(incremental);
     self
   }
 
+  /// Set whether to enable top level await.
   pub fn top_level_await(&mut self, top_level_await: bool) -> &mut Self {
     self.top_level_await = Some(top_level_await);
     self
   }
 
+  /// Set the cache options.
   pub fn cache(&mut self, cache: ExperimentCacheOptions) -> &mut Self {
     self.cache = Some(cache);
     self
   }
 
+  /// Set whether to enable future defaults.
   pub fn future_defaults(&mut self, future_defaults: bool) -> &mut Self {
     self.future_defaults = Some(future_defaults);
     self
   }
 
+  /// Set whether to enable css.
   pub fn css(&mut self, css: bool) -> &mut Self {
     self.css = Some(css);
     self
   }
 
+  /// Set whether to enable async web assembly.
   pub fn async_web_assembly(&mut self, async_web_assembly: bool) -> &mut Self {
     self.async_web_assembly = Some(async_web_assembly);
     self
   }
 
+  /// Set whether to enable parallel code splitting.
   pub fn parallel_code_splitting(&mut self, parallel_code_splitting: bool) -> &mut Self {
     self.parallel_code_splitting = Some(parallel_code_splitting);
     self
   }
 
+  /// Build [`Experiments`] from options.
+  ///
+  /// [`Experiments`]: rspack_core::options::Experiments
   pub fn build(
     &mut self,
     _builder_context: &mut BuilderContext,
@@ -3245,6 +3446,8 @@ impl ExperimentsBuilder {
 
 #[cfg(test)]
 mod test {
+  use enum_tag::EnumTag;
+
   use super::*;
 
   #[test]
@@ -3295,5 +3498,22 @@ mod test {
       .module(ModuleOptionsBuilder::default().no_parse(ModuleNoParseRules::Rules(vec![])))
       .node(NodeOptionBuilder::default().dirname(NodeDirnameOption::EvalOnly))
       .build(&mut Default::default());
+  }
+
+  #[test]
+  #[should_panic]
+  #[allow(unreachable_code, unused_variables)]
+  fn use_options_directly() {
+    let optimization: Optimization = todo!();
+    let output: OutputOptions = todo!();
+    let experiments: Experiments = todo!();
+    let module: ModuleOptions = todo!();
+    let node: Option<NodeOption> = todo!();
+    Compiler::builder()
+      .optimization(optimization)
+      .output(output)
+      .experiments(experiments)
+      .module(module)
+      .node(node);
   }
 }
