+++ crates/rspack/src/lib.rs
@@ -1,1 +1,5 @@
+#![forbid(missing_docs)]
+
+//! The Rspack compiler.
+
 pub mod builder;
